db_service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==9.1.0
